Add a site.pp file and the provisioner script
Previously we committed a Vagrantfile that used the shell provisioner
to ensure that Puppet was installed on our node...and I forgot to commit
the script I used to DO that.  Also, there's a site.pp file used for
classification, and I forgot to commit THAT.  This commit fixes both
of those issues.